IsTreeSymmetric
Finds if a binary tree is symmetric. Code works by passing the two sides of the tree that are to be checked for mirror images. But since one side would be left and the other right, we need to check both left and right subtrees, we accomplish this in recursion using:
                            IsMirror(LeftRoot->Left, RightRoot->Right) && IsMirror(LeftRoot->Right, RightRoot->Left);